# library/internal/server/models.go
package server

// RequestAuthor - author info
type RequestAuthor struct {
	FullName  string `json:"full_name"`
	BirthDate string `json:"birth_date,omitempty"`
	DeathDate string `json:"death_date,omitempty"`
}

type RequestAuthorWithID struct {
	Id        string `json:"id"`
	FullName  string `json:"full_name"`
	BirthDate string `json:"birth_date,omitempty"`
	DeathDate string `json:"death_date,omitempty"`
}

type ResponseAuthorShortInfo struct {
	FullName string `json:"full_name"`
	Id       string `json:"id"`
}

type ResponseAuthorFullInfo struct {
	FullName  string `json:"full_name"`
	BirthDate string `json:"birth_date,omitempty"`
	DeathDate string `json:"death_date,omitempty"`
}

type ResponseBook struct {
	Id    string `json:"id"`
	Title string `json:"title"`
}

type RequestBook struct {
	Title   string   `json:"title"`
	Authors []string `json:"authors"`
}

type RequestBookWithID struct {
	Id      string   `json:"id"`
	Title   string   `json:"title"`
	Authors []string `json:"authors"`
}

type ResponseBookFullInfo struct {
	Title   string   `json:"title"`
	Authors []string `json:"authors"`
}

type ErrorResponse struct {
	Error string `json:"error"`
}